@@ -78,45 +26,59 @@
   },
   "homepage": "https://github.com/jcoreio/async-throttle#readme",
   "devDependencies": {
-    "@babel/cli": "^7.12.10",
-    "@babel/core": "^7.12.10",
     "@babel/eslint-parser": "^7.12.13",
     "@babel/node": "^7.20.7",
-    "@babel/plugin-transform-runtime": "^7.12.10",
-    "@babel/preset-env": "^7.12.11",
-    "@babel/preset-flow": "^7.12.1",
-    "@babel/register": "^7.12.10",
-    "@commitlint/cli": "^11.0.0",
-    "@commitlint/config-conventional": "^11.0.0",
-    "@jedwards1211/commitlint-config": "^1.0.2",
-    "@jedwards1211/eslint-config": "^2.0.2",
-    "@jedwards1211/eslint-config-flow": "^3.0.0",
-    "babel-eslint": "^10.1.0",
-    "babel-plugin-istanbul": "^6.0.0",
-    "chai": "^4.2.0",
+    "@babel/plugin-syntax-flow": "^7.14.5",
+    "@babel/plugin-transform-react-jsx": "^7.14.9",
+    "@jcoreio/eslint-plugin-implicit-dependencies": "^1.1.1",
+    "@jcoreio/toolchain": "^4.9.1",
+    "@jcoreio/toolchain-circle": "^4.9.1",
+    "@jcoreio/toolchain-esnext": "^4.9.1",
+    "@jcoreio/toolchain-flow": "^4.9.1",
+    "@jcoreio/toolchain-mocha": "^4.9.1",
+    "@jcoreio/toolchain-semantic-release": "^4.9.1",
+    "@jcoreio/toolchain-typescript": "^4.9.1",
+    "@types/chai": "^4.3.5",
+    "@types/chai-as-promised": "^7.1.5",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^22.7.4",
+    "@types/sinon": "^9.0.3",
+    "@typescript-eslint/eslint-plugin": "^7.6.0",
+    "@typescript-eslint/parser": "^7.6.0",
+    "chai": "^4.3.7",
     "chai-as-promised": "^7.1.1",
-    "codecov": "^3.8.1",
     "copy": "^0.3.2",
     "cross-env": "^7.0.3",
-    "eslint": "^7.18.0",
+    "eslint": "^8.56.0",
     "eslint-config-prettier": "^7.2.0",
-    "eslint-plugin-flowtype": "^5.2.0",
+    "eslint-plugin-flowtype": "^8.0.3",
+    "eslint-plugin-no-only-tests": "^3.1.0",
     "flow-bin": "^0.143.1",
-    "flow-copy-source": "https://github.com/jedwards1211/flow-copy-source#no-watch",
-    "husky": "^4.3.8",
-    "istanbul": "^0.4.5",
-    "lint-staged": "^10.5.3",
-    "mocha": "^8.2.1",
-    "nyc": "^15.1.0",
-    "prettier": "^2.2.1",
-    "prettier-eslint": "^12.0.0",
+    "mocha": "^10.2.0",
     "rimraf": "^3.0.2",
-    "semantic-release": "^17.3.7",
     "sinon": "^9.0.3",
-    "typescript": "^4.1.3",
+    "typescript": "^5.1.0",
     "waait": "^1.0.5"
   },
   "dependencies": {
-    "@babel/runtime": "^7.12.5"
+    "@babel/runtime": "^7.18.6"
+  },
+  "main": "dist/index.js",
+  "module": "dist/index.mjs",
+  "types": "dist/index.d.ts",
+  "exports": {
+    "./package.json": "./package.json",
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.mjs",
+      "default": "./dist/index.js"
+    }
+  },
+  "engines": {
+    "node": ">=16"
+  },
+  "packageManager": "pnpm@8.11.0",
+  "@jcoreio/toolchain": {
+    "migratedVersion": "4.9.1"
   }
 }
